changes: wrappers and replacing lib with helpers
