+++ frontend/__tests__/src/pages/Chapters.test.tsx
@@ -1,13 +1,13 @@
 import { fireEvent, screen, waitFor } from '@testing-library/react'
 
+import { fetchAlgoliaData } from 'helpers/api'
+import { render } from 'helpers/wrappers/test-util'
 import { useNavigate } from 'react-router-dom'
-import { fetchAlgoliaData } from 'lib/api'
-import { render } from 'lib/test-util'
 
 import ChaptersPage from 'pages/Chapters'
 import { mockChapterData } from '@tests/data/mockChapterData'
 
-jest.mock('lib/api', () => ({
+jest.mock('helpers/api', () => ({
   fetchAlgoliaData: jest.fn(),
 }))
 jest.mock('react-router-dom', () => ({
