+++ frontend/components.json
@@ -14,7 +14,7 @@
     "components": "components",
     "utils": "utils",
     "ui": "components/ui",
-    "lib": "lib",
+    "lib": "helpers",
     "hooks": "hooks"
   },
   "iconLibrary": "lucide"
