+++ frontend/src/components/Card.tsx
@@ -1,10 +1,10 @@
 import { FontAwesomeIcon, FontAwesomeIconProps } from '@fortawesome/react-fontawesome'
+import { CardProps, tooltipStyle } from 'helpers/constants'
+import FontAwesomeIconWrapper from 'helpers/wrappers/FontAwesomeIconWrapper'
 import { useState, useEffect } from 'react'
 import { Tooltip } from 'react-tooltip'
 import { Icons } from 'utils/data'
-import { CardProps, tooltipStyle } from 'lib/constants'
-import FontAwesomeIconWrapper from 'lib/FontAwesomeIconWrapper'
-import { cn } from 'lib/utils'
+import { cn } from 'utils/utils'
 import ActionButton from 'components/ActionButton'
 import ContributorAvatar from 'components/ContributorAvatar'
 import DisplayIcon from 'components/DisplayIcon'
