+++ frontend/src/components/ContributorAvatar.tsx
@@ -1,6 +1,5 @@
+import { tooltipStyle, topContributorsType } from 'helpers/constants'
 import { Tooltip } from 'react-tooltip'
-
-import { tooltipStyle, topContributorsType } from 'lib/constants'
 
 const ContributorAvatar = ({ contributor }: { contributor: topContributorsType }) => {
   return (
