+++ frontend/src/components/Footer.tsx
@@ -1,6 +1,6 @@
+import { footerSections, Section } from 'helpers/constants'
 import { ChevronDown, ChevronUp } from 'lucide-react'
 import { useState, useCallback } from 'react'
-import { footerSections, Section } from 'utils/constants'
 
 export default function Footer() {
   // State to keep track of the open section in the footer
