+++ frontend/src/components/TopicBadge.tsx
@@ -1,6 +1,5 @@
+import { tooltipStyle } from 'helpers/constants'
 import { Tooltip } from 'react-tooltip'
-
-import { tooltipStyle } from 'lib/constants'
 
 const TopicBadge = ({
   topic,
