+++ frontend/src/helpers/wrappers/FontAwesomeIconWrapper.tsx
